we're really getting down to it, here.
notes: after the presentation tomorrow, this will be moved off of mongo,
and probably off of express as well, using just a tiny internal server
and something along the lines of minimongo, miniredis, or even just
indexeddb. that'll bring the size of binaries down hopefully under 50
megs, at which point it'll actually make for a viable consumer app.